Revert "add logs chat id"
This reverts commit a8e87c6b5f9d2407a76b870c58c75dcc274811c5.

+++ src/main/kotlin/com/github/insanusmokrassar/TimingPostsTelegramBot/plugins/builtin/GarbageCollector.kt
@@ -66,7 +66,7 @@ class GarbageCollector(
         if (dataPair.second < config.minimalRate) {
             deletePost(
                 bot,
-                baseConfig.sourceChatId,
+                baseConfig.sourceChatId.toLong(),
                 dataPair.first
             )
         }
